fix: rename component RegistrySearches to History
